# chore: optimize sync manga filter