Performance of `filterFavoritesAndNonFavorites` on my 512 mangas library with Android Studio connected:

Before:

- 24 seconds
- 24 seconds
- 23 seconds

After:

- 5 seconds
- 4 seconds
- 4 seconds